refactor(renderer): move CameraController out of Renderer
