+++ src/@types/typedArrayBuffer.d.ts
@@ -1,0 +1,5 @@
+type TypedArrayBuffer =
+Int8Array | Int16Array | Int32Array |
+Uint8ClampedArray |
+Uint8Array | Uint16Array | Uint32Array
+Float32Array | Float64Array;
